# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: BodyLogic - AI-Powered Health & Fitness Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/messages:
    get:
      operationId: getAiMessages
      tags: [ai]
      summary: Get AI coach conversation history
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message to the AI coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAiMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"
...
components:
  schemas:
...
    AiMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, role, content, createdAt]
    SendAiMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...